requirement.txt
python-multipart
fastapi 
uvicorn 
google-generativeai 
pillow
